backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu
Pillow>=10.0.0
python-multipart==0.0.9
numpy>=1.24.0
